requirements.txt
backports.entry-points-selectable==1.1.1
cachelib==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
filelock==3.4.0
Flask==2.0.2
Flask-Mobility==1.1.0
Flask-ReCaptcha==0.4.2
Flask-Session==0.4.0
Flask-WTF==1.0.0
fonttools==4.28.5
fpdf==1.7.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
numpy==1.22.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
platformdirs==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
virtualenv==20.10.0
Werkzeug==2.0.2
WTForms==3.0.1
